Fixed reviews bugs and updated readme

+++ web/benefactors/models.py
@@ -46,6 +46,8 @@ def load_user(user_id):
     return User.query.get(int(user_id))
 
 
+# -----------------------------------------------------Account----------------------------------------------------------
+
 class User(db.Model, UserMixin):
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(20), unique=True, nullable=False)
@@ -80,6 +82,8 @@ class User(db.Model, UserMixin):
         return f"User('{self.username}', '{self.email}', '{self.user_image}')"
 
 
+# ----------------------------------------------------Posts-------------------------------------------------------------
+
 class Post(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(100), nullable=False)
@@ -94,6 +98,8 @@ class Post(db.Model):
         return f"Post('{self.title}', '{self.date_posted}')"
 
 
+# ----------------------------------------------------Comments----------------------------------------------------------
+
 class PostComment(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     comment_desc = db.Column(db.Text, nullable=False)
@@ -105,17 +111,21 @@ class PostComment(db.Model):
         return f"PostComment('{self.post_id}', '{self.user_id}', '{self.comment_desc}')"
 
 
+# ----------------------------------------------------Reviews-----------------------------------------------------------
+
 class UserReview(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     description = db.Column(db.Text)
     date_posted = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    score = db.Column(db.Integer, nullable=False)
+    score = db.Column(db.Numeric, nullable=False)
     author = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     profile = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
 
     def __repr__(self):
         return f"UserReview('{self.description}', '{self.score}''{self.date_posted}')"
 
+
+# ------------------------------------------------------Messages--------------------------------------------------------
 
 # User1 and User2 cannot switch position, let's say user1 has a chat channel with user2, user 2 should have the same channel with user 1.
 # You cannot have two channels between two same users twice, unless the channel is closed
@@ -128,7 +138,7 @@ class ChatChannel(db.Model):
     last_updated = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
     user1_status = db.Column(db.Enum(channelStatusEnum), default=channelStatusEnum.READ)
     user2_status = db.Column(db.Enum(channelStatusEnum), default=channelStatusEnum.READ)
-    
+
     user_1 = db.relationship("User", backref=backref("usr_1", uselist=False), foreign_keys=[user1_id])
     user_2 = db.relationship("User", backref=backref("usr_2", uselist=False), foreign_keys=[user2_id])
